feat(chart): add proxy config override template
This adds a template for proxy-config Secret which is meant
to replace the proxy-config secret in the osm subchart. Because
the default value of .Values.fluentBit.enableProxySupport in the
osm subchart is false AND we do not change this value as part of the
parent chart, we can create a new proxy-config Secret in the parent
chart if the azureHTTPProxy and azureHTTPSProxy values are passed in
at install time.

Signed-off-by: Michelle Noorali <minooral@microsoft.com>